Fix a deadlock bug.
The system previously sent LAST_BATCH_SENT_FLAG only to predecessors of a step that had sent a last batch to output.
i.e. the step should be done, but only the predecessors get told to quit.

When a load stage ends with a step that has multiple replicas, it will wait for the next stage, but
not send a signal to tell the other replicas to quit working.

I added sending a signal to the step that sent the last batch to output so that all the replicas are
closed if needed.