Made some changes to fix scheduling LMs in different stages.
Realized there may be an issue with routing batches to steps that are in different load stages,
upon testing, this did deadlock. I tracked down a few fundamental issues and made changes to allow
batches to go to different stages. See the comments next to the changes for detailed explanations
on the causes of the issues and the solution.