Replace old caching with batch and LM input/output caching.
I am moving to a batch level caching system because the pipeline caching and resuming was frail. With changes I had made and even from before I forked, the cache system fails to resume in a lot of cases. I also don't see how it can be modular as it attempts to be (sharing steps between pipelines for instance).

I considered a simple state of the pipeline cache, which would be pretty effective but totally unaccomadating to changes and not modular, but fast and cheap on disk space,
vs batch caching which is somewhat modular and tolerant of changes to pipeline structure, sharing across pipelines, etc. It is still pretty sensitive due to batching maybe putting different rows in batches (explained in the README), but it should be pretty good. A backup is the LM level caching. Batch level caching re-runs the pipeline so it is kind of slow, but it is e.g. 10 min so not too bad.
Thus, I chose batc caching.